requirements.txt
asgiref==3.2.7
aspy.yaml==1.3.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.9.11
cffi==1.13.2
cfgv==2.0.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
decorator==4.4.1
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-cors-headers==3.2.1
django-cors-middleware==1.5.0
django-environ==0.4.5
django-extensions==2.2.9
django-oauth2-provider==0.2.6.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-storages==1.9.1
django-templated-mail==1.1.1
django-test-plus==1.4.0
django-url-checks==0.2.0
django-username-email==2.2.4
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
entrypoints==0.3
environ==1.0
factory-boy==2.12.0
Faker==4.0.2
flake8==3.7.9
flake8-bugbear==20.1.4
flake8-docstrings==1.5.0
flake8-polyfill==1.0.2
google-api-core==1.16.0
google-auth==1.12.0
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
identify==1.4.7
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
nodeenv==1.3.3
notebook==6.0.2
oauthlib==3.1.0
pandocfilters==1.4.2
parso==0.5.1
pep8-naming==0.10.0
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
pre-commit==2.2.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.4.2
PyJWT==1.7.1
pyrsistent==0.15.5
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
Send2Trash==1.5.0
shortuuid==0.5.0
six==1.13.0
snowballstemmer==2.0.0
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
toml==0.10.0
tornado==6.0.3
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.7
virtualenv==16.7.8
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==5.0.1
widgetsnbextension==3.5.1
zipp==0.6.0
